# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'tazavera_api API Documentation'
  description: 'REST API for the Tazavera specialty coffee platform. Provides endpoints for managing coffee offerings, sensory evaluations (cupping), user profiles, and supporting taxonomies. Authentication is handled via Laravel Passport (OAuth 2.0) using the Authorization Code flow with PKCE for first-party clients and the Password grant for trusted clients.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: "This API uses **Laravel Passport (OAuth 2.0)**. Authenticated endpoints require a Bearer access token in the `Authorization` header. Two grant flows are supported:\n\n**Authorization Code + PKCE** (recommended, for public/first-party clients such as SPAs and mobile apps that cannot safely store a client secret):\n1. Generate a `code_verifier` (random string) and derive a `code_challenge` = BASE64URL(SHA256(`code_verifier`)).\n2. Redirect the user to `GET /oauth/authorize` with `response_type=code`, your `client_id`, `redirect_uri`, `scope`, `state`, `code_challenge`, and `code_challenge_method=S256`.\n3. After the user approves, exchange the returned `code` at `POST /oauth/token` with `grant_type=authorization_code`, the `code`, `redirect_uri`, `client_id`, and the original `code_verifier`.\n4. The response contains `access_token`, `refresh_token`, and `expires_in`.\n\n**Password grant** (for trusted first-party clients only):\n- `POST /oauth/token` with `grant_type=password`, `client_id`, `client_secret`, the user's `username` and `password`, and the required `scope`.\n- The response contains `access_token`, `refresh_token`, and `expires_in`.\n\nSend the token as: `Authorization: Bearer {ACCESS_TOKEN}`.\n\n**Scopes:** endpoints enforce `profile:read` and/or `profile:write`. Request the appropriate scope when obtaining your token. When not scope parameter scope is requested, it will be defaulted to 'profile:red', in the case an unappropiate scope is requested, the authorization will be rejected."
security:
  -
    default: []
paths:
...
  '/users/{user_ulid}/password':
    put:
      summary: 'Update user password'
      operationId: updateUserPassword
      description: "Changes a user's password. The current password must be supplied and is\nverified before the change is applied.\n\n**Authorization:** requires the `profile:write` scope and permission to\nupdate this user (policy). A non-owner receives `403 Forbidden`."
      parameters: []
      responses:
        200:
          description: 'Password updated'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Password updated successfully.'
                properties:
                  message:
                    type: string
                    example: 'Password updated successfully.'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'This action is unauthorized.'
                properties:
                  message:
                    type: string
                    example: 'This action is unauthorized.'
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  description: "The user's current password. Required for verification."
                  example: OldP@ss123
                password:
                  type: string
                  description: 'The new password. Must meet default strength rules and match passwordConfirmation. The value and <code>passwordConfirmation</code> must match.'
                  example: NewStr0ng!
                passwordConfirmation:
                  type: string
                  description: 'Confirmation of the new password. Must match the password field. Note: camelCase, not the Laravel-default password_confirmation.'
                  example: NewStr0ng!
              required:
                - currentPassword
                - password
                - passwordConfirmation
    parameters:
      -
        in: path
        name: user_ulid
        description: 'The ULID of the user.'
        example: 01M35F5RX4ADGC3CSDXXYB08DA
        required: true
        schema:
          type: string
...